Refactoring:
The goal was to make the tests in Holes.hs pass. The Part data type
is extended to include now the complete prefix (in the Middle and
Last variant). findOperators is replaced by a findBoth that gives
priority to continuing parts before beginning parts.

Although the user interface of the parser is just an operator
table, the shunt function should work on a part table.

A lot of function should be renamed.