Fix the last test.
The last test introduced in the previous commit was failing.
It was testing that

    f f do
      f

was parsed similary as

    f f do
           f

but the Indent module was forcing an indented block to start to the
right of the introducing keyword (here `do`).

With this restriction lifted, this behaves similarly to Haskell, e.g.

    f = do
      f

happens all the time.